+++ src/app/api/profile/update/route.js
@@ -2,7 +2,46 @@ import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 import dbConnect from "@/lib/db.connect";
 import { collectionNamesObj } from "@/lib/db.connect";
+import { writeFile, mkdir } from "fs/promises";
+import { join } from "path";
 
+// GET - 
+export async function GET(request) {
+  try {
+    const session = await getServerSession();
+    
+    if (!session || !session.user) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    const usersCollection = await dbConnect(collectionNamesObj.usersCollection);
+    const user = await usersCollection.findOne({ email: session.user.email });
+
+    if (!user) {
+      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const profileData = {
+      name: user.name || "",
+      email: user.email,
+      bio: user.bio || "",
+      skills: user.skills || [],
+      contactInfo: user.contactInfo || {},
+      profileImage: user.profileImage || "",
+      role: user.role || "user",
+      createdAt: user.createdAt,
+      updatedAt: user.updatedAt
+    };
+
+    return NextResponse.json(profileData);
+
+  } catch (error) {
+    console.error("Profile fetch error:", error);
+    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
+  }
+}
+
+// PUT - 
 export async function PUT(request) {
   try {
     const session = await getServerSession();
@@ -47,7 +86,8 @@ export async function PUT(request) {
   }
 }
 
-export async function GET(request) {
+// POST - 
+export async function POST(request) {
   try {
     const session = await getServerSession();
     
@@ -55,29 +95,77 @@ export async function GET(request) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    const usersCollection = await dbConnect(collectionNamesObj.usersCollection);
-    const user = await usersCollection.findOne({ email: session.user.email });
+    const formData = await request.formData();
+    const imageFile = formData.get('profileImage');
 
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    if (!imageFile) {
+      return NextResponse.json({ error: "No image file provided" }, { status: 400 });
     }
 
-    const profileData = {
-      name: user.name || "",
-      email: user.email,
-      bio: user.bio || "",
-      skills: user.skills || [],
-      contactInfo: user.contactInfo || {},
-      profileImage: user.profileImage || "",
-      role: user.role || "user",
-      createdAt: user.createdAt,
-      updatedAt: user.updatedAt
-    };
+    // Validate file type
+    if (!imageFile.type.startsWith('image/')) {
+      return NextResponse.json({ error: "Please select a valid image file" }, { status: 400 });
+    }
 
-    return NextResponse.json(profileData);
+    // Validate file size (max 5MB)
+    if (imageFile.size > 5 * 1024 * 1024) {
+      return NextResponse.json({ error: "Image size should be less than 5MB" }, { status: 400 });
+    }
+
+    // Convert File to Buffer
+    const bytes = await imageFile.arrayBuffer();
+    const buffer = Buffer.from(bytes);
+
+    // Create uploads directory if it doesn't exist
+    const uploadsDir = join(process.cwd(), 'public/uploads/profiles');
+    try {
+      await mkdir(uploadsDir, { recursive: true });
+    } catch (error) {
+      console.error('Error creating directory:', error);
+      return NextResponse.json({ error: "Server configuration error" }, { status: 500 });
+    }
+
+    // Generate unique filename
+    const timestamp = Date.now();
+    const filename = `${session.user.email}-${timestamp}.jpg`;
+    const filepath = join(uploadsDir, filename);
+
+    // Save file to filesystem
+    try {
+      await writeFile(filepath, buffer);
+    } catch (error) {
+      console.error('Error saving file:', error);
+      return NextResponse.json({ error: "Error saving image file" }, { status: 500 });
+    }
+
+    // Image URL that will be accessible from the browser
+    const imageUrl = `/uploads/profiles/${filename}`;
+
+    // Save image URL to database
+    const usersCollection = await dbConnect(collectionNamesObj.usersCollection);
+    
+    const result = await usersCollection.updateOne(
+      { email: session.user.email },
+      { 
+        $set: { 
+          profileImage: imageUrl,
+          updatedAt: new Date()
+        } 
+      }
+    );
+
+    if (result.modifiedCount === 0) {
+      return NextResponse.json({ error: "Profile image update failed" }, { status: 400 });
+    }
+
+    return NextResponse.json({ 
+      success: true, 
+      imageUrl: imageUrl,
+      message: "Profile image updated successfully" 
+    });
 
   } catch (error) {
-    console.error("Profile fetch error:", error);
+    console.error("Profile image upload error:", error);
     return NextResponse.json({ error: "Internal server error" }, { status: 500 });
   }
 }
